+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Migrated next_prompt on WhatsApp messages to buttons
-- Removed next_prompt on WhatsApp messages
+
+### Deprecated
+- next_prompt on WhatsApp messages is deprecated, use buttons instead. Will be removed on the next major release
 
 ## v1.1.0-dev.3